requirements.txt
streamlit>=1.44.0
torch>=2.6.0
torchvision>=0.21.0
torchaudio>=2.6.0
scikit-learn>=1.6.0
pandas>=2.2.0
numpy>=2.2.0
matplotlib>=3.10.0
seaborn>=0.13.0
Pillow>=11.0.0
PyYAML>=6.0.2
tqdm>=4.67.0
